Rework config gen again.
